+++ ft-sdk/src/main/java/com/ft/sdk/garble/bean/ActionBean.java
@@ -91,6 +91,17 @@ public class ActionBean {
         return property;
     }
 
+
+    HashMap<String, Object> tags = new HashMap<>();
+
+    public HashMap<String, Object> getTags() {
+        return tags;
+    }
+
+    public void setTags(HashMap<String, Object> tags) {
+        this.tags = tags;
+    }
+
     public ActionBean() {
 
     }
@@ -217,6 +228,7 @@ public class ActionBean {
     public String getAttrJsonString() {
         HashMap<String, Object> map = new HashMap<>();
         map.put(Constants.KEY_RUM_PROPERTY, property);
+        map.put(Constants.KEY_RUM_TAGS, tags);
         map.put(Constants.KEY_HAS_REPLAY, hasReplay);
         return Utils.hashMapObjectToJson(map);
     }
@@ -240,6 +252,16 @@ public class ActionBean {
                     this.property.put(key, jsonProperty.opt(key));
                 }
             }
+
+            JSONObject jsonTags = json.optJSONObject(Constants.KEY_RUM_TAGS);
+            if (jsonTags != null) {
+                Iterator<String> keys = jsonTags.keys();
+                while (keys.hasNext()) {
+                    String key = keys.next();
+                    this.tags.put(key, jsonTags.opt(key));
+                }
+            }
+
             this.hasReplay = json.optBoolean(Constants.KEY_HAS_REPLAY, false);
         } catch (JSONException e) {
             LogUtils.e(TAG, LogUtils.getStackTraceString(e));