+++ ft-sdk/src/main/java/com/ft/sdk/garble/bean/ViewBean.java
@@ -145,6 +145,16 @@ public class ViewBean {
         return property;
     }
 
+    HashMap<String, Object> tags = new HashMap<>();
+
+    public HashMap<String, Object> getTags() {
+        return tags;
+    }
+
+    public void setTags(HashMap<String, Object> tags) {
+        this.tags = tags;
+    }
+
     public String getViewReferrer() {
         return viewReferrer;
     }
@@ -339,6 +349,7 @@ public class ViewBean {
         map.put(Constants.KEY_MEMORY_MAX, memoryMax);
         map.put(Constants.KEY_HAS_REPLAY, hasReplay);
         map.put(Constants.KEY_RUM_PROPERTY, property);
+        map.put(Constants.KEY_RUM_TAGS, tags);
         return Utils.hashMapObjectToJson(map);
     }
 
@@ -367,6 +378,16 @@ public class ViewBean {
                 while (keys.hasNext()) {
                     String key = keys.next();
                     this.property.put(key, jsonProperty.opt(key));
+                }
+            }
+
+            JSONObject jsonTags = json.optJSONObject(Constants.KEY_RUM_TAGS);
+            if (jsonTags != null) {
+                Iterator<String> keys = jsonTags.keys();
+
+                while (keys.hasNext()) {
+                    String key = keys.next();
+                    this.tags.put(key, jsonTags.opt(key));
                 }
             }
         } catch (JSONException e) {